Convert fetchPermissions.js to TS
